Watch live index collisions and entry key16 collisions
lotta print spam, but points out TT collisions (i.e. a simultaneous collision of both index and key16) when they occur.

Even at `Hash value 80000`, a tt collision is reliably produced around depth 40 or so (lesser depths with smaller tables).

Subjectively, it seems that index collisions have leading bits in common between the two colliding key64s.
Subjectively, it seems that as table size goes up (and therefore index collision probability goes down), the leading bit commonality increases.
(That is, with tables on the order of 1-100 MB, digits in common is around 3-5; on the order of 10-80 GB, it's 8-9 digits in common. Subjectively, dont @ me)